backend/requirements.txt
fastapi
uvicorn
pymongo>=4.6.0
motor>=3.3.0
bcrypt
pyjwt
python-dotenv
google-cloud-dialogflow
email-validator
jinja2==3.1.2